+++ code/src/cljs/sixsq/nuvla/ui/edge_detail/subs.cljs
@@ -27,6 +27,13 @@
 
 
 (reg-sub
+  ::nuvlabox-components
+  :<- [::nuvlabox-status]
+  (fn [{:keys [components]}]
+    components))
+
+
+(reg-sub
   ::nuvlabox-vulns
   (fn [db]
     (::spec/nuvlabox-vulns db)))