+++ CHANGELOG.md
@@ -4,11 +4,14 @@
 
 ### Added
 
+- NuvlaBox detail page - add Logs tab
+- Resource log reusable component
 - Main component - ResponsiveMenuBar
 - Session/Profile - Two factor authentication TOTP method support
 
 ### Changed
 
+- Fixed "stop logs" action button for deployment and NuvlaBox logs
 - Edge detail - Use ResponsiveMenuBar
 - Cimi detail - Use ResponsiveMenuBar
 - Cimi detail - Format operation simplified and transformed in a plain function